requirements.txt
aiohappyeyeballs==2.4.8
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==3.3.2
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.115.11
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.43
langchain-deepseek==0.1.2
langchain-huggingface==0.1.2
langchain-openai==0.3.8
langchain-pinecone==0.2.3
langchain-tests==0.3.13
langchain-text-splitters==0.3.6
langsmith==0.3.11
llama-cloud==0.1.13
llama-cloud-services==0.6.3
llama-index==0.12.22
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.22
llama-index-embeddings-huggingface==0.4.0
llama-index-embeddings-huggingface-optimum==0.3.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.25
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.6
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-pinecone==0.4.4
llama-parse==0.6.2
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
openai==1.65.4
optimum==1.24.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pluggy==1.5.0
propcache==0.3.0
protobuf==6.30.0
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pypdf==5.3.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
striprtf==0.0.26
sympy==1.13.1
syrupy==4.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
